hp calculation added

+++ backend/src/main/java/com/utn/javaproject/dndsheets/domain/dto/CharacterStatsDto.java
@@ -21,5 +21,5 @@ public class CharacterStatsDto {
     private HashMap<String,Short> abilityScores;
     private List<Long> velocities;
     private HashMap<String, Short> proficiencies;
-    private Long hp;
+    private Integer hp;
 }
